Spostato gestione birre nel domain layer e piccolo refactoring
