+++ app/src/main/java/it/unimib/brewday/ui/gestisci_ricette/CreaRicettaFragment.java
@@ -30,7 +30,7 @@ import it.unimib.brewday.model.Ingrediente;
 import it.unimib.brewday.model.IngredienteRicetta;
 import it.unimib.brewday.model.Ricetta;
 import it.unimib.brewday.ui.gestisci_ingredienti.AdapterListViewIngredienti;
-import it.unimib.brewday.util.GestioneRicette;
+import it.unimib.brewday.domain.GestioneRicette;
 import it.unimib.brewday.util.ListaIngredienti;
 
 public class CreaRicettaFragment extends Fragment {
